updated manger for new structured db
Started with structured and versions. Structured is added to the base registry name from the XML file. It should prevent older dbs from mucking with the newer if all goes well. At least for harvest as harvest has to take the name and convert it to an ENUM. The manager tools are more flexible, but it should cause most some pain in getting around the structured suffix.

Then added versioning. The tool now builds a version table and checks its own version against what the DB wants as a minimum. Verified its operation. The create-registry will skip versioning checks because the table has not been created yet.