update date to string
Part 1 was to fix the configuration. Had to make same change in harvest. Really should have only one data-dic-types.cfg in registry-common and not two want to be the same in regsistry-mgr and harvest.

Part 2: ditch the data-dic-data.jar because it is old and defines dates as opensearch date. Need to rebuild it with new code on this branch.

Part 3: fix LDD loader to set index correctly.